REFACTOR: remove necessity of commons to implementations
