+++ library-tv/build.gradle
@@ -65,7 +65,6 @@ dependencies {
     implementation rootProject.androidDependencies.kotlin
     implementation rootProject.androidDependencies.appCompat
     implementation rootProject.androidDependencies.leanback
-    implementation rootProject.androidDependencies.commons
 
     testImplementation rootProject.testDependencies.junit
     testImplementation rootProject.testDependencies.robolectric
